#50 refactor: actor-settings の store/context を _stores/actor-settings/ へ統合
- _stores/actor-settings-store.ts と _contexts/actor-settings-store-context.tsx を _stores/actor-settings/{store.ts,context.tsx,index.ts} にまとめる
- context.tsx は共通生成関数 createStoreContext 経由に置き換え
- store と対応 context を同一ディレクトリに配置し依存関係を分かりやすくする
