+++ src/prototypes/time-control/time-control-03/__tests__/intent-store.test.ts
@@ -1,4 +1,4 @@
-import { createActorSettingsStore } from '../_stores/actor-settings-store'
+import { createActorSettingsStore } from '../_stores/actor-settings'
 import { createActorStore } from '../_stores/actor-store'
 import { createGameClockStore } from '../_stores/game-clock-store'
 import { createIntentStore } from '../_stores/intent-store'
