+++ src/prototypes/time-control/time-control-03/_components/action-bar/index.tsx
@@ -1,12 +1,14 @@
 import { Button } from '@/components/ui/button'
 
-import { useActorSettingsStore } from '../../_contexts/actor-settings-store-context'
 import { useActorStore } from '../../_contexts/actor-store-context'
 import { useGameClockStore } from '../../_contexts/game-clock-store-context'
 import { usePathStore } from '../../_contexts/path-store-context'
 import { usePlannedPathStore } from '../../_contexts/planned-path-store-context'
 import { usePositionStore } from '../../_contexts/position-store-context'
-import { DEFAULT_ACTOR_SETTINGS } from '../../_stores/actor-settings-store'
+import {
+  DEFAULT_ACTOR_SETTINGS,
+  useActorSettingsStore,
+} from '../../_stores/actor-settings'
 import { ActorId } from '../../types'
 
 type ActionBarProps = {
