+++ src/prototypes/time-control/time-control-03/_contexts/stores-provider.tsx
@@ -1,6 +1,9 @@
 import { ReactNode, useState } from 'react'
 
-import { createActorSettingsStore } from '../_stores/actor-settings-store'
+import {
+  ActorSettingsStoreContext,
+  createActorSettingsStore,
+} from '../_stores/actor-settings'
 import { createActorStore } from '../_stores/actor-store'
 import { createGameClockStore } from '../_stores/game-clock-store'
 import { createIntentStore } from '../_stores/intent-store'
@@ -8,7 +11,6 @@ import { createPathStore } from '../_stores/path-store'
 import { createPlannedPathStore } from '../_stores/planned-path-store'
 import { createPositionStore } from '../_stores/position-store'
 
-import { ActorSettingsStoreContext } from './actor-settings-store-context'
 import { ActorStoreContext } from './actor-store-context'
 import { GameClockStoreContext } from './game-clock-store-context'
 import { IntentStoreContext } from './intent-store-context'
