+++ src/prototypes/time-control/time-control-03/_stores/actor-settings/context.tsx
@@ -1,0 +1,9 @@
+import { createStoreContext } from '@/stores/utils/create-store-context'
+
+import { ActorSettingsState } from './store'
+
+const { StoreContext, useStoreSelector } =
+  createStoreContext<ActorSettingsState>('ActorSettings')
+
+export const ActorSettingsStoreContext = StoreContext
+export const useActorSettingsStore = useStoreSelector
