+++ src/prototypes/time-control/time-control-03/_stores/position-store.ts
@@ -3,10 +3,7 @@ import { createStore, StoreApi } from 'zustand/vanilla'
 import { getTickMs } from '../../time-control-02/_lib/get-tick-ms'
 import { ActionLogEntry, ActorId, Position } from '../types'
 
-import {
-  ActorSettingsStore,
-  DEFAULT_ACTOR_SETTINGS,
-} from './actor-settings-store'
+import { ActorSettingsStore, DEFAULT_ACTOR_SETTINGS } from './actor-settings'
 import { ActorStore, DEFAULT_ACTOR_INFO } from './actor-store'
 import { GameClockStore } from './game-clock-store'
 import { PathStore } from './path-store'
@@ -42,7 +39,7 @@ const INITIAL_STATE = {
 
 /**
  * @param actorStore 移動速度・tick 発生頻度 (`_stores/actor-store.ts`)
- * @param actorSettingsStore 行動進行モード (`_stores/actor-settings-store.ts`)
+ * @param actorSettingsStore 行動進行モード (`_stores/actor-settings/store.ts`)
  * @param pathStore 残り移動経路 (`_stores/path-store.ts`)
  * @param gameClockStore 共通ゲームクロック・履歴ログ (`_stores/game-clock-store.ts`)
  */
